refactor: delete legacy evaluator classes; score_fns are the source of truth
The legacy evaluator surface (rllm/eval/evaluator/*.py + class definitions
in rllm/eval/types.py) was a parallel grading path that had become
redundant after the score_fn migration. Remove it.

Changes:
- Delete rllm/eval/evaluator/ entirely (bfcl, ifeval, llm_equality,
  llm_judge, translation, widesearch). Their logic moves into the
  matching rllm/eval/score_fns/<name>.py module so score_fns are
  self-contained.
- Delete the 9 legacy evaluator classes from rllm/eval/types.py
  (MathEvaluator, CountdownEvaluator, CodeEvaluator, F1Evaluator,
  MCQEvaluator, IoUEvaluator, PointInMaskEvaluator, DepthEvaluator,
  CompoundEvaluator). The protocols (AgentFlow, Evaluator, EvalOutput,
  Signal, Task, AgentConfig) and _extract_agent_answer helper stay.
- evaluator_loader._EVALUATOR_REGISTRY now maps names to score_fn
  import paths. _load_and_instantiate wraps bare functions in
  _FunctionEvaluator so legacy callers (EvalRunner, agenthub agents)
  passing `task: dict` still work — the wrapper auto-converts to
  rllm.task.Task before invoking the score_fn.
- Update tests/eval/test_evaluator_loader.py to assert protocol
  conformance rather than specific class identity. Adds a test for
  the dict-call backward-compat path.
- rllm/eval/__init__.py drops CompoundEvaluator export.

Verified:
- rllm eval gsm8k --max-examples 3 → 100%
- rllm eval ./harbor/examples/tasks/hello-world --sandbox-backend docker --agent bash → 100%
- All 14 reward-fn names resolve via load_evaluator.
- Legacy `evaluator.evaluate(task: dict, episode)` call still works.
