--- rllm/eval/evaluator/translation.py
@@ -1,90 +1,0 @@
-"""Translation evaluator using ChrF (character n-gram F-score).
-
-ChrF is the standard metric for machine translation evaluation that works
-well across languages without tokenization. Falls back to LLM judge if
-configured (for XCOMET-style evaluation).
-"""
-
-from __future__ import annotations
-
-import logging
-from collections import Counter
-
-from rllm.eval.types import EvalOutput, Signal, _extract_agent_answer
-from rllm.types import Episode
-
-logger = logging.getLogger(__name__)
-
-
-class TranslationEvaluator:
-    """Evaluator for translation tasks using ChrF score.
-
-    ChrF computes character n-gram precision and recall between hypothesis
-    and reference, then takes the F-score. This is language-agnostic and
-    the standard WMT metric alongside BLEU.
-    """
-
-    def __init__(self, max_n: int = 6, beta: float = 2.0):
-        self.max_n = max_n
-        self.beta = beta
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        hypothesis = _extract_agent_answer(episode)
-        reference = str(task.get("ground_truth", ""))
-
-        if not reference:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="chrf", value=0.0)],
-                metadata={"reason": "no_reference"},
-            )
-
-        score = self._compute_chrf(hypothesis, reference)
-        is_correct = score >= 0.5
-        return EvalOutput(
-            reward=score,
-            is_correct=is_correct,
-            signals=[Signal(name="chrf", value=score)],
-        )
-
-    def _char_ngrams(self, text: str, n: int) -> Counter:
-        """Extract character n-grams from text."""
-        return Counter(text[i : i + n] for i in range(len(text) - n + 1))
-
-    def _compute_chrf(self, hypothesis: str, reference: str) -> float:
-        """Compute ChrF score between hypothesis and reference."""
-        if not hypothesis or not reference:
-            return 0.0
-
-        total_precision = 0.0
-        total_recall = 0.0
-        count = 0
-
-        for n in range(1, self.max_n + 1):
-            hyp_ngrams = self._char_ngrams(hypothesis, n)
-            ref_ngrams = self._char_ngrams(reference, n)
-
-            if not hyp_ngrams or not ref_ngrams:
-                continue
-
-            common = sum((hyp_ngrams & ref_ngrams).values())
-            precision = common / sum(hyp_ngrams.values()) if sum(hyp_ngrams.values()) > 0 else 0.0
-            recall = common / sum(ref_ngrams.values()) if sum(ref_ngrams.values()) > 0 else 0.0
-
-            total_precision += precision
-            total_recall += recall
-            count += 1
-
-        if count == 0:
-            return 0.0
-
-        avg_precision = total_precision / count
-        avg_recall = total_recall / count
-
-        if avg_precision + avg_recall == 0:
-            return 0.0
-
-        beta_sq = self.beta**2
-        chrf = (1 + beta_sq) * avg_precision * avg_recall / (beta_sq * avg_precision + avg_recall)
-        return chrf
