+++ rllm/eval/types.py
@@ -135,432 +135,3 @@ def _extract_agent_answer(episode: Episode) -> str:
         if traj.steps:
             return str(traj.steps[-1].output or "")
     return ""
-
-
-# ---------------------------------------------------------------------------
-# Built-in evaluators
-# ---------------------------------------------------------------------------
-
-
-class MathEvaluator:
-    """Evaluator for math tasks using extract_answer + grade_answer from math_utils."""
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        from rllm.rewards.math_utils.utils import extract_answer, grade_answer_mathd, grade_answer_sympy
-
-        answer_text = _extract_agent_answer(episode)
-
-        # Extract model answer from boxed notation
-        model_answer = extract_answer(answer_text)
-        if model_answer is None:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="accuracy", value=0.0)],
-                metadata={"reason": "no_answer_extracted"},
-            )
-
-        # Get ground truth(s)
-        ground_truths = task.get("ground_truth")
-        if ground_truths is None:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="accuracy", value=0.0)],
-                metadata={"reason": "no_ground_truth"},
-            )
-
-        if isinstance(ground_truths, str | float | int):
-            ground_truths = [ground_truths]
-
-        # Process ground truths (extract from boxed if needed)
-        processed = []
-        for truth in ground_truths:
-            truth = str(truth)
-            if "\\boxed" in truth:
-                extracted = extract_answer(truth)
-                if extracted is not None:
-                    processed.append(extracted)
-            else:
-                processed.append(truth)
-
-        if not processed:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="accuracy", value=0.0)],
-                metadata={"reason": "no_processed_ground_truth"},
-            )
-
-        # Grade against all possible correct answers
-        for ground_truth in processed:
-            is_correct = grade_answer_mathd(model_answer, ground_truth) or grade_answer_sympy(model_answer, ground_truth)
-            if is_correct:
-                return EvalOutput(
-                    reward=1.0,
-                    is_correct=True,
-                    signals=[Signal(name="accuracy", value=1.0)],
-                )
-
-        return EvalOutput(
-            reward=0.0,
-            is_correct=False,
-            signals=[Signal(name="accuracy", value=0.0)],
-        )
-
-
-class CountdownEvaluator:
-    """Evaluator for countdown arithmetic tasks."""
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        from rllm.rewards.countdown_reward import compute_score
-
-        answer_text = _extract_agent_answer(episode)
-        target = task.get("target")
-        nums = task.get("nums", [])
-
-        if target is None or not nums:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="accuracy", value=0.0)],
-                metadata={"reason": "missing_target_or_nums"},
-            )
-
-        ground_truth = {"target": target, "numbers": nums}
-        score = compute_score(answer_text, ground_truth)
-
-        is_correct = score >= 1.0
-        reward = 1.0 if is_correct else 0.0
-        return EvalOutput(
-            reward=reward,
-            is_correct=is_correct,
-            signals=[Signal(name="accuracy", value=float(is_correct))],
-        )
-
-
-class CodeEvaluator:
-    """Evaluator for code generation tasks using code execution."""
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        from rllm.rewards.code_reward import RewardCodeFn
-        from rllm.rewards.reward_types import RewardConfig
-
-        answer_text = _extract_agent_answer(episode)
-        reward_fn = RewardCodeFn(RewardConfig())
-        reward_output = reward_fn(task_info=task, action=answer_text)
-
-        is_correct = reward_output.reward > 0
-        return EvalOutput(
-            reward=float(reward_output.reward),
-            is_correct=is_correct,
-            signals=[Signal(name="accuracy", value=1.0 if is_correct else 0.0)],
-            metadata=reward_output.metadata,
-        )
-
-
-class F1Evaluator:
-    """Evaluator using token-overlap F1 score."""
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        import re
-        import string
-        from collections import Counter
-
-        answer_text = _extract_agent_answer(episode)
-        gold_text = task.get("ground_truth", "")
-        if gold_text is None:
-            gold_text = ""
-
-        def normalize_text(s: str) -> str:
-            s = s.lower()
-            s = "".join(ch for ch in s if ch not in set(string.punctuation))
-            s = re.sub(r"\b(a|an|the)\b", " ", s)
-            return " ".join(s.split())
-
-        pred_norm = normalize_text(str(answer_text))
-        gold_norm = normalize_text(str(gold_text))
-        pred_tokens = pred_norm.split()
-        gold_tokens = gold_norm.split()
-
-        if not pred_tokens or not gold_tokens:
-            f1 = 0.0
-        else:
-            common = Counter(pred_tokens) & Counter(gold_tokens)
-            num_same = sum(common.values())
-            if num_same == 0:
-                f1 = 0.0
-            else:
-                precision = num_same / len(pred_tokens)
-                recall = num_same / len(gold_tokens)
-                f1 = (2 * precision * recall) / (precision + recall) if (precision + recall) > 0 else 0.0
-
-        is_correct = f1 > 0
-        return EvalOutput(
-            reward=f1,
-            is_correct=is_correct,
-            signals=[Signal(name="f1", value=f1)],
-        )
-
-
-class MCQEvaluator:
-    """Evaluator for multiple-choice question tasks.
-
-    Extracts the choice letter (A-J) from the model response and compares
-    against the ground_truth field.
-    """
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        answer_text = _extract_agent_answer(episode)
-        model_letter = self._extract_choice_letter(answer_text)
-
-        gt = task.get("ground_truth", "")
-        expected_letter = str(gt).strip().upper()[:1] if gt else ""
-
-        is_correct = model_letter != "" and model_letter == expected_letter
-        reward = 1.0 if is_correct else 0.0
-        return EvalOutput(
-            reward=reward,
-            is_correct=is_correct,
-            signals=[Signal(name="accuracy", value=reward)],
-            metadata={"model_answer": model_letter, "expected": expected_letter},
-        )
-
-    @staticmethod
-    def _extract_choice_letter(text: str) -> str:
-        """Extract the choice letter (A-J) from model response."""
-        import re
-
-        text = text.strip()
-        if not text:
-            return ""
-        # Try exact single letter
-        if len(text) == 1 and text.upper() in "ABCDEFGHIJ":
-            return text.upper()
-        # Try pattern like "The answer is (B)" or "Answer: C"
-        match = re.search(r"(?:answer\s*(?:is|:)\s*\(?([A-Ja-j])\)?)", text, re.IGNORECASE)
-        if match:
-            return match.group(1).upper()
-        # Try pattern like "**B**" or "(B)"
-        match = re.search(r"(?:\*\*([A-J])\*\*|\(([A-J])\))", text)
-        if match:
-            return (match.group(1) or match.group(2)).upper()
-        # Fallback: first standalone capital letter A-J
-        match = re.search(r"\b([A-J])\b", text)
-        if match:
-            return match.group(1)
-        return ""
-
-
-class IoUEvaluator:
-    """Evaluator for visual grounding tasks using Intersection-over-Union.
-
-    Parses ``[x1, y1, x2, y2]`` from model output and computes IoU against
-    the ground truth bounding box. Returns reward 1.0 if IoU >= 0.5.
-    """
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        answer_text = _extract_agent_answer(episode)
-        pred_bbox = self._parse_bbox(answer_text)
-        gt_bbox = task.get("ground_truth_bbox", task.get("ground_truth"))
-
-        if pred_bbox is None or gt_bbox is None:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="iou", value=0.0)],
-                metadata={"reason": "parse_failure"},
-            )
-
-        if isinstance(gt_bbox, str):
-            gt_bbox = self._parse_bbox(gt_bbox)
-        if isinstance(gt_bbox, list | tuple) and len(gt_bbox) == 4:
-            gt_bbox = [float(x) for x in gt_bbox]
-        else:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="iou", value=0.0)],
-                metadata={"reason": "invalid_ground_truth"},
-            )
-
-        iou = self._compute_iou(pred_bbox, gt_bbox)
-        is_correct = iou >= 0.5
-        return EvalOutput(
-            reward=1.0 if is_correct else 0.0,
-            is_correct=is_correct,
-            signals=[Signal(name="iou", value=iou)],
-        )
-
-    @staticmethod
-    def _parse_bbox(text: str) -> list[float] | None:
-        import re
-
-        match = re.search(r"\[?\s*(\d+(?:\.\d+)?)\s*[,\s]\s*(\d+(?:\.\d+)?)\s*[,\s]\s*(\d+(?:\.\d+)?)\s*[,\s]\s*(\d+(?:\.\d+)?)\s*\]?", text)
-        if match:
-            return [float(match.group(i)) for i in range(1, 5)]
-        return None
-
-    @staticmethod
-    def _compute_iou(box_a: list[float], box_b: list[float]) -> float:
-        x1 = max(box_a[0], box_b[0])
-        y1 = max(box_a[1], box_b[1])
-        x2 = min(box_a[2], box_b[2])
-        y2 = min(box_a[3], box_b[3])
-        inter = max(0, x2 - x1) * max(0, y2 - y1)
-        area_a = max(0, box_a[2] - box_a[0]) * max(0, box_a[3] - box_a[1])
-        area_b = max(0, box_b[2] - box_b[0]) * max(0, box_b[3] - box_b[1])
-        union = area_a + area_b - inter
-        return inter / union if union > 0 else 0.0
-
-
-class PointInMaskEvaluator:
-    """Evaluator for spatial reasoning tasks with point-in-mask checking.
-
-    Parses ``(x, y)`` point coordinates from model output and checks if the
-    predicted point falls within the ground truth mask region.
-    """
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        answer_text = _extract_agent_answer(episode)
-        point = self._parse_point(answer_text)
-
-        if point is None:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="point_accuracy", value=0.0)],
-                metadata={"reason": "parse_failure"},
-            )
-
-        mask_data = task.get("ground_truth_mask")
-        if mask_data is None:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="point_accuracy", value=0.0)],
-                metadata={"reason": "no_mask"},
-            )
-
-        try:
-            is_in_mask = self._check_point_in_mask(point, mask_data)
-        except Exception:
-            is_in_mask = False
-
-        reward = 1.0 if is_in_mask else 0.0
-        return EvalOutput(
-            reward=reward,
-            is_correct=is_in_mask,
-            signals=[Signal(name="point_accuracy", value=reward)],
-        )
-
-    @staticmethod
-    def _parse_point(text: str) -> tuple[float, float] | None:
-        import re
-
-        match = re.search(r"\(?\s*(\d+(?:\.\d+)?)\s*[,\s]\s*(\d+(?:\.\d+)?)\s*\)?", text)
-        if match:
-            return (float(match.group(1)), float(match.group(2)))
-        return None
-
-    @staticmethod
-    def _check_point_in_mask(point: tuple[float, float], mask_data: bytes) -> bool:
-        import io
-
-        from PIL import Image
-
-        if isinstance(mask_data, bytes):
-            img = Image.open(io.BytesIO(mask_data))
-        else:
-            img = mask_data
-        img = img.convert("L")
-
-        x, y = point
-        w, h = img.size
-        px = int(x * w / 1000) if x > 1 else int(x * w)
-        py = int(y * h / 1000) if y > 1 else int(y * h)
-        px = max(0, min(px, w - 1))
-        py = max(0, min(py, h - 1))
-
-        return img.getpixel((px, py)) > 127
-
-
-class DepthEvaluator:
-    """Evaluator for depth estimation tasks using absolute relative error.
-
-    Parses a depth value from model output and computes
-    ``AbsRel = |pred - gt| / gt``. Returns reward = ``max(0, 1 - absrel)``.
-    """
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        answer_text = _extract_agent_answer(episode)
-        pred_depth = self._parse_depth(answer_text)
-
-        gt_depth_raw = task.get("ground_truth", "")
-        try:
-            gt_depth = float(gt_depth_raw)
-        except (TypeError, ValueError):
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="absrel", value=1.0)],
-                metadata={"reason": "invalid_ground_truth"},
-            )
-
-        if pred_depth is None or gt_depth <= 0:
-            return EvalOutput(
-                reward=0.0,
-                is_correct=False,
-                signals=[Signal(name="absrel", value=1.0)],
-                metadata={"reason": "parse_failure"},
-            )
-
-        absrel = abs(pred_depth - gt_depth) / gt_depth
-        reward = max(0.0, 1.0 - absrel)
-        is_correct = reward > 0.5
-        return EvalOutput(
-            reward=reward,
-            is_correct=is_correct,
-            signals=[Signal(name="absrel", value=absrel)],
-        )
-
-    @staticmethod
-    def _parse_depth(text: str) -> float | None:
-        import re
-
-        match = re.search(r"(\d+(?:\.\d+)?)\s*(?:m|meters|metre)?", text)
-        if match:
-            return float(match.group(1))
-        return None
-
-
-class CompoundEvaluator:
-    """Runs multiple evaluators and merges their signals.
-
-    The reward is a weighted average of individual evaluator rewards.
-    """
-
-    def __init__(self, evaluators: list[tuple[Evaluator, float]]):
-        """Initialize with list of (evaluator, weight) tuples."""
-        self.evaluators = evaluators
-
-    def evaluate(self, task: dict, episode: Episode) -> EvalOutput:
-        all_signals: list[Signal] = []
-        weighted_reward = 0.0
-        total_weight = 0.0
-        any_correct = False
-
-        for evaluator, weight in self.evaluators:
-            output = evaluator.evaluate(task, episode)
-            weighted_reward += output.reward * weight
-            total_weight += weight
-            all_signals.extend(output.signals)
-            if output.is_correct:
-                any_correct = True
-
-        reward = weighted_reward / total_weight if total_weight > 0 else 0.0
-        return EvalOutput(
-            reward=reward,
-            is_correct=any_correct,
-            signals=all_signals,
-        )
